requirements.txt
openai>=1.12.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.26.2
